requirements.txt
docopt==0.6.2
numpy==2.0.2
pykwalify==1.8.0
pyradiomics @ file:///Users/phd/Documents/pyradiomics
python-dateutil==2.9.0.post0
PyWavelets==1.6.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
simpleitk==2.5.2
six==1.17.0
